+++ src/main/kotlin/at/voyagers/common/account/application/AccountWriter.kt
@@ -1,9 +1,8 @@
-package at.voyagers.common.account.business.component
+package at.voyagers.common.account.application
 
 import at.voyagers.common.account.domain.Account
 import at.voyagers.common.account.domain.AccountRepository
 import at.voyagers.common.account.domain.AccountRole
-import at.voyagers.common.account.domain.CreateAccountSpec
 import org.springframework.stereotype.Component
 
 @Component
